+++ src/test/java/io/justedlev/msrvs/configcenter/ConfigCenterApplicationTests.java
@@ -1,4 +1,4 @@
-package io.justedlev.msrvs.cinfigcenter;
+package io.justedlev.msrvs.configcenter;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
